task 8

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -4,21 +4,20 @@
 -- Note: An average score can be a decimal
 DELIMITER //
 
-CREATE PROCEDURE ComputeAverageScoreForUser(
-    IN user_id INT
-)
+CREATE PROCEDURE ComputeAverageScoreForUser(IN user_id INT)
 BEGIN
-    DECLARE avg_score DECIMAL(10, 2);
+    DECLARE avg_score FLOAT;
 
-    -- Compute the average score for the user
+    -- Calculate the average score for the given user_id
     SELECT AVG(score) INTO avg_score
     FROM corrections
     WHERE user_id = user_id;
 
-    -- Update the average score in the users table
+    -- Update the average_score in the users table
     UPDATE users
     SET average_score = avg_score
     WHERE id = user_id;
 END //
 
 DELIMITER ;
+
